mesa pixman xserver git update 13 Feb 2013
xserver          commit 7115f6c709898a5124b67e19c61dc01334471358
pixman           commit 5e207f825bd1ed3142a623bcbceca00508907c5e
mesa             commit 8cabe26f5dc4c66ae59649cfd7b25f761cda250b

+++ pixman/pixman/pixman-sse2.c
@@ -4558,7 +4558,7 @@ sse2_composite_add_n_8888 (pixman_implementation_t *imp,
 	dst = dst_line;
 	dst_line += dst_stride;
 
-	while (w && (unsigned long)dst & 15)
+	while (w && (uintptr_t)dst & 15)
 	{
 	    d = *dst;
 	    *dst++ =
@@ -4617,7 +4617,7 @@ sse2_composite_add_n_8_8888 (pixman_implementation_t *imp,
 	mask_line += mask_stride;
 	w = width;
 
-	while (w && ((unsigned long)dst & 15))
+	while (w && ((uintptr_t)dst & 15))
 	{
 	    uint8_t m = *mask++;
 	    if (m)
